refactoring nearly complete
- reset of toggle with real toggled key not working
