--- tap_threads.py
@@ -1,5 +1,0 @@
-    
-# from threading import Thread
-# from time import sleep
-# from tap_keyboard import Key_Event
-# import pygetwindow as gw # to get name of actual window for focusapp function
